% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the ShinyApp}
\usage{
run_app(mae)
}
\arguments{
\item{mae}{A MultiAssayExperiment maeect (for documentation purposes only).}
}
\description{
Launches a Shiny application using the provided MultiAssayExperiment maeect.
}
\examples{
\dontrun{
# Load required packages

library(MultiAssayExperiment)
library(ShinyCITExpresso)
# Load your MultiAssayExperiment maeect (example)
data("example_mae")

# Run the ShinyApp (not run in documentation)
run_app(mae = example_mae)
}
}
