% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the ShinyApp}
\usage{
run_app(mae)
}
\arguments{
\item{mae}{A MultiAssayExperiment object (for documentation purposes only).}
}
\description{
Launches a Shiny application using the provided MultiAssayExperiment maeect.
}
\examples{
\dontrun{
run_app(mae = mae)
}

}
